requirements.txt
fastapi==0.111.0
uvicorn
pypdf
python-dotenv
langchain
langchain-huggingface
faiss-cpu
huggingface_hub
requests
